+++ lib/src/frames/frames.dart
@@ -3,7 +3,7 @@ export 'comment_frame.dart';
 export 'frame.dart';
 export 'frame_parser.dart';
 export 'picture_frame.dart';
-//export 'table_of_contents_frame.dart';
+export 'table_of_contents_frame.dart';
 export 'text_information_frame.dart';
 //export 'url_frame.dart';
 export 'user_url_frame.dart';